--- twake/frontend/src/app/features/debug/types/debug-recoil-state-types.ts
@@ -1,5 +1,0 @@
-export interface TwakeDebugState {
-  dumpStateSnapshot?(): void;
-  get?(key: string): void;
-  getAllAtoms?(): void;
-}
